+++ src/hooks/queries/useUserQuery.tsx
@@ -5,6 +5,11 @@ import { AxiosError, AxiosResponse } from "axios";
 import { useMutation } from "react-query";
 import { encrypt } from "../useCrypto";
 import { getCheckNicknameApi } from "@/apis/getCheckNicknameApi";
+import { useModalStore } from "@/stores/modalStore";
+import { useNavigate } from "react-router-dom";
+import { PATH } from "@/constants/path";
+import CommonModal from "@/components/Common/CommonModal/CommonModal";
+import CommonMutationErrorModal from "@/components/Error/CommonMutationErrorModal/CommonMutationErrorModal";
 
 interface PostSignUpMutateType {
   identifier: string;
@@ -12,10 +17,10 @@ interface PostSignUpMutateType {
   information: string;
   interest: CheckboxValueType[];
 }
-interface PostSignUpType {
-  onError: () => void;
-}
-export const usePostSignUp = ({ onError }: PostSignUpType) => {
+
+export const usePostSignUp = () => {
+  const { setModal, closeModal } = useModalStore();
+  const navigate = useNavigate();
   const { mutate, isLoading, mutateAsync, data } = useMutation(
     ({ identifier, nickname, information, interest }: PostSignUpMutateType) =>
       signUpApi({ identifier, nickname, information, interest }),
@@ -25,7 +30,17 @@ export const usePostSignUp = ({ onError }: PostSignUpType) => {
         localStorage.setItem(IDENTIFIER, encrypt(identifier));
       },
       onError: () => {
-        onError();
+        const onClickMoveToSiupUpFirstStep = () => {
+          closeModal();
+          navigate(PATH.LOGIN);
+        };
+        setModal(
+          <CommonModal
+            content={"오류가 발생했습니다.\n처음으로 이동합니다."}
+            buttonContent="확인"
+            onClick={onClickMoveToSiupUpFirstStep}
+          />
+        );
       },
     }
   );
@@ -37,24 +52,27 @@ interface GetCheckNinkNameMutateType {
 }
 interface GetCheckNinkNameType {
   onSuccess: (res: AxiosResponse) => void;
-  onError: (err: AxiosError<{ message?: string }>) => void;
+  onError: () => void;
 }
 
 export const useGetCheckNickName = ({
   onSuccess,
   onError,
 }: GetCheckNinkNameType) => {
-  const { mutate } = useMutation(
+  const { setModal, closeModal } = useModalStore();
+  const { mutate, isLoading } = useMutation(
     ({ value }: GetCheckNinkNameMutateType) => getCheckNicknameApi({ value }),
     {
       onSuccess: (res: AxiosResponse) => {
         onSuccess(res);
       },
       onError: (err: AxiosError<{ message?: string }>) => {
-        onError(err);
+        setModal(
+          <CommonMutationErrorModal error={err} closeModal={closeModal} />
+        );
+        onError();
       },
-      useErrorBoundary: false,
     }
   );
-  return { mutate };
+  return { mutate, isLoading };
 };
